+++ sofa-ark-parent/core-impl/container/src/main/java/com/alipay/sofa/ark/container/model/BizModel.java
@@ -42,7 +42,10 @@ import com.alipay.sofa.ark.spi.service.biz.BizManagerService;
 import com.alipay.sofa.ark.spi.service.event.EventAdminService;
 
 import java.io.File;
+import java.io.IOException;
 import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
@@ -52,7 +55,8 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-import static com.alipay.sofa.ark.api.ArkClient.recycleBizTempWorkDir;
+import static com.alipay.sofa.ark.spi.constant.Constants.BIZ_TEMP_WORK_DIR_RECYCLE_FILE_SUFFIX;
+import static org.apache.commons.io.FileUtils.deleteQuietly;
 
 /**
  * Ark Biz Standard Model
@@ -522,5 +526,57 @@ public class BizModel implements Biz {
             }
         }
         return false;
+    }
+
+    /**
+     * recycle biz temp work dir
+     *
+     * @param bizTempWorkDir
+     * @return
+     */
+    public static boolean recycleBizTempWorkDir(File bizTempWorkDir) {
+        if (bizTempWorkDir == null) {
+            return false;
+        }
+
+        if (bizTempWorkDir.isDirectory()) {
+            try {
+                String newPath = markBizTempWorkDirRecycled(bizTempWorkDir);
+                File markedFile = new File(newPath);
+                if (!markedFile.exists()) {
+                    ArkLoggerFactory.getDefaultLogger().warn(
+                        String.format("when delete marked biz temp work dir %s, file not exists ",
+                            markedFile.getPath()));
+                    return false;
+                }
+
+                return deleteQuietly(markedFile);
+            } catch (IOException e) {
+                throw new ArkRuntimeException("mark and delete biz temp work dir error: "
+                                              + e.getMessage());
+            }
+
+        }
+
+        return deleteQuietly(bizTempWorkDir);
+    }
+
+    /**
+     * mark biz temp work dir is recycled
+     *
+     * @param bizTempWorkDir
+     * @return
+     * @throws IOException
+     */
+    private static String markBizTempWorkDirRecycled(File bizTempWorkDir) throws IOException {
+        String sourcePath = bizTempWorkDir.getAbsolutePath();
+        String targetPath = String.format("%s-%s-%s", sourcePath, System.currentTimeMillis(),
+            BIZ_TEMP_WORK_DIR_RECYCLE_FILE_SUFFIX);
+
+        Files.move(Paths.get(sourcePath), Paths.get(targetPath));
+        ArkLoggerFactory.getDefaultLogger().info(
+            String.format("move biz temp work dir from %s to %s", sourcePath, targetPath));
+
+        return targetPath;
     }
 }
