+++ sofa-ark-parent/core-impl/container/src/test/java/com/alipay/sofa/ark/container/model/BizModelTest.java
@@ -20,16 +20,18 @@ import com.alipay.sofa.ark.spi.model.BizInfo.BizStateRecord;
 import com.alipay.sofa.ark.spi.model.BizState;
 import org.junit.Test;
 
+import java.io.File;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
+import static com.alipay.sofa.ark.api.ArkConfigs.getStringValue;
+import static com.alipay.sofa.ark.api.ArkConfigs.init;
+import static java.util.Arrays.asList;
+import static org.apache.commons.io.FileUtils.touch;
+import static org.junit.Assert.*;
 
 public class BizModelTest {
 
@@ -85,5 +87,33 @@ public class BizModelTest {
         assertTrue(bizModel.toString().contains("-> activated"));
         assertTrue(bizModel.toString().contains("-> deactivated"));
         assertTrue(bizModel.toString().contains("-> unresolved"));
+    }
+
+    @Test
+    public void testLoadConfigFile() throws Exception {
+        URL resource = this.getClass().getClassLoader().getResource("test.props");
+        init(asList(resource));
+        assertEquals("b123", getStringValue("a123"));
+        assertEquals("d123", getStringValue("c123"));
+    }
+
+    @Test
+    public void testRecycleBizTempWorkDir() throws Throwable {
+        assertFalse(BizModel.recycleBizTempWorkDir(null));
+
+        File fileJar = new File("/tmp/" + System.currentTimeMillis() + ".jar");
+        touch(fileJar);
+
+        assertTrue(BizModel.recycleBizTempWorkDir(fileJar));
+        assertFalse(fileJar.exists());
+
+        File fileDir = new File("/tmp/" + System.currentTimeMillis() + "-test");
+        fileDir.mkdir();
+        File fileSubFile = new File(fileDir.getAbsolutePath() + "/subfile.jar");
+        touch(fileSubFile);
+
+        assertTrue(BizModel.recycleBizTempWorkDir(fileDir));
+        assertFalse(fileDir.exists());
+        assertFalse(fileSubFile.exists());
     }
 }
