+++ sofa-ark-parent/core/api/src/main/java/com/alipay/sofa/ark/api/ArkClient.java
@@ -425,56 +425,4 @@ public class ArkClient {
     private static ArkLogger getLogger() {
         return ArkLoggerFactory.getDefaultLogger();
     }
-
-    /**
-     * recycle biz temp work dir
-     *
-     * @param bizTempWorkDir
-     * @return
-     */
-    public static boolean recycleBizTempWorkDir(File bizTempWorkDir) {
-        if (bizTempWorkDir == null) {
-            return false;
-        }
-
-        if (bizTempWorkDir.isDirectory()) {
-            try {
-                String newPath = markBizTempWorkDirRecycled(bizTempWorkDir);
-                File markedFile = new File(newPath);
-                if (!markedFile.exists()) {
-                    getLogger().warn(
-                        String.format("when delete marked biz temp work dir %s, file not exists ",
-                            markedFile.getPath()));
-                    return false;
-                }
-
-                return deleteQuietly(markedFile);
-            } catch (IOException e) {
-                throw new ArkRuntimeException("mark and delete biz temp work dir error: "
-                                              + e.getMessage());
-            }
-
-        }
-
-        return deleteQuietly(bizTempWorkDir);
-    }
-
-    /**
-     * mark biz temp work dir is recycled
-     *
-     * @param bizTempWorkDir
-     * @return
-     * @throws IOException
-     */
-    private static String markBizTempWorkDirRecycled(File bizTempWorkDir) throws IOException {
-        String sourcePath = bizTempWorkDir.getAbsolutePath();
-        String targetPath = String.format("%s-%s-%s", sourcePath, System.currentTimeMillis(),
-            BIZ_TEMP_WORK_DIR_RECYCLE_FILE_SUFFIX);
-
-        Files.move(Paths.get(sourcePath), Paths.get(targetPath));
-        getLogger().info(
-            String.format("move biz temp work dir from %s to %s", sourcePath, targetPath));
-
-        return targetPath;
-    }
 }
